requirements.txt
pepperoni
sqlalchemy
cx_Oracle
paramiko
flask
flask_httpauth
waitress
sqlparse
gitpython
